Patches by cras: unsigned int's to unsigned short in packet 	routines (bugs in HPUX). Removed config.h's from contrib's 	getopt.c

+++ lib/silcske/payload_internal.h
@@ -24,7 +24,7 @@
 /* SILC Key Exchange Start Payload */
 typedef struct {
   unsigned char flags;
-  unsigned int len;
+  unsigned short len;
 
   unsigned char *cookie;
   unsigned short cookie_len;
